refactor: rename folder conponent to components
